refactor: migrate FareMapService from tentative service class
